+++ src/rle_ctx.c
@@ -595,12 +595,12 @@ void rle_ctx_dump(struct rle_ctx_management *_this, struct rle_configuration *rl
 	PRINT("\tlabel_type		\t= [0x%0x]\n", _this->label_type);
 	PRINT("\tend address		\t= [%p]\n", _this->end_address);
 	PRINT("\tLink Status:\n");
-	PRINT("\tPackets sent/received	\t= [%lu]\n", (long unsigned int)_this->lk_status.counter_ok);
-	PRINT("\tPackets lost		\t= [%lu]\n", (long unsigned int)_this->lk_status.counter_lost);
-	PRINT("\tPackets dropped	\t\t= [%lu]\n",
-	      (long unsigned int)_this->lk_status.counter_dropped);
-	PRINT("\tBytes sent/received	\t= [%lu]\n",
-	      (long unsigned int)_this->lk_status.counter_bytes_ok);
+	PRINT("\tPackets sent/received	\t= [%llu]\n", (long long unsigned int)_this->lk_status.counter_ok);
+	PRINT("\tPackets lost		\t= [%llu]\n", (long long unsigned int)_this->lk_status.counter_lost);
+	PRINT("\tPackets dropped	\t\t= [%llu]\n",
+	      (long long unsigned int)_this->lk_status.counter_dropped);
+	PRINT("\tBytes sent/received	\t= [%llu]\n",
+	      (long long unsigned int)_this->lk_status.counter_bytes_ok);
 
 	if (_this->frag_counter == 0) {
 		PRINT("\n--------------------------------------------------\n");
